+++ shared/core/networkutil/src/commonMain/kotlin/com/thomaskioko/tvmaniac/core/networkutil/Failure.kt
@@ -1,0 +1,12 @@
+package com.thomaskioko.tvmaniac.core.networkutil
+
+
+sealed class Failure(
+    val throwable: Throwable,
+    val errorMessage: String = ""
+)
+
+class DefaultError(val message: String) : Failure(
+    throwable = Throwable(message),
+    errorMessage = message
+)
